README: refactor project structure section
This PR fixes an issue where "High-order components" was incorrectly
used.

In fact, this is a side-effect of this PR, which refactors the section
"Estrutura do projeto" so it provides more information about each
directory and why they exists.

This also fixes references from "a Pilot" to "o Pilot", since Pilot
refers to "the project", which is male in Portuguese -- "O projeto
Pilot".